+++ bindings/src/roboplan/visualization.py
@@ -375,3 +375,18 @@ def visualizePositionTrace(
             color=waypoint_color,
             position=positions[idx],
         )
+
+
+def se3_to_viser_wxyz(transform: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
+    """Extract Viser-compatible position and wxyz quaternion from an SE(3) matrix.
+
+    Args:
+        transform: 4x4 homogeneous transform.
+
+    Returns:
+        Tuple of ``(position, wxyz)`` as numpy arrays.
+    """
+    position = transform[:3, 3].copy()
+    quat = pin.Quaternion(transform[:3, :3])
+    wxyz = np.array([quat.w, quat.x, quat.y, quat.z])
+    return position, wxyz
